requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.14.2
cachetools==5.3.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.19.1
flatbuffers==24.3.25
fonttools==4.59.2
frozendict==2.4.6
fsspec==2025.9.0
gast==0.4.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos
grpcio==1.64.1
h11==0.16.0
h5py==3.11.0
httplib2==0.31.0
huggingface-hub==0.35.3
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.0.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mplfinance==0.12.10b0
mpmath==1.3.0
multitasking==0.0.12
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.18.2
pillow==11.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus
protobuf
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
pywin32==306
PyYAML==6.0.3
pyzmq==26.0.3
regex==2025.9.18
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.13.1
selenium==4.34.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.22.1
torch==2.8.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yfinance==0.2.66
youtube-transcript-api==0.6.2
zipp==3.19.2
schedule
gunicorn
psycopg2-binary
